feat: Introduce ProductDetailResultModel and its use case (#140)
This commit introduces `ProductDetailResultModel` to encapsulate product details along with additional information like the number of pending reservation requests if the current user is the product owner.

**Key changes:**
-   **`ProductDetailResultModel.kt`**: New data class containing `ProductDetailModel` and an optional `requestCount` (Int).
-   **`GetProductDetailResultUseCase.kt`**:
    -   New use case that fetches product details.
    -   Maps category ID to category name.
    -   Checks if the authenticated user is the product owner.
    -   If the user is the owner, it calculates the number of pending reservation requests (`RentalStatus.PENDING`).
    -   Returns a `Result<ProductDetailResultModel>`.
-   **`GetProductDetailUseCase.kt`**: Deleted, as its functionality is now incorporated into `GetProductDetailResultUseCase`.
