+++ app/src/main/java/com/example/rentit/domain/product/model/ProductDetailResultModel.kt
@@ -1,0 +1,6 @@
+package com.example.rentit.domain.product.model
+
+data class ProductDetailResultModel(
+    val productDetail: ProductDetailModel,
+    val requestCount: Int? = null
+)
